requirements.txt
aiogram==3.13.0
requests==2.32.3
python-dotenv==1.0.1
tiktok_downloader==0.3.5
yt_dlp==2024.8.6
pymysql==1.1.1
instagrapi==2.1.2
pillow==10.4.0
